compose_example/backend/requirements.txt
fastapi[all]
uvicorn[standard]
gunicorn
strawberry-graphql
aiohttp